strawberry_api/requirements.txt
aiosqlite==0.20.0
fastapi==0.110.1
fastapi-cors
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
strawberry-graphql[fastapi]==0.224.1
uvicorn==0.29.0
